requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
sqlalchemy
pyjwt
passlib[bcrypt]